requirements.txt
flask
flask_sqlalchemy
python-decouple